+++ examples/vue3/src/main.js
@@ -1,5 +1,5 @@
 import { createApp } from 'vue'
-import VueAnimXyz from '@animxyz/vue/src/vue3'
+import VueAnimXyz from '@animxyz/vue3'
 import App from './App.vue'
 
 // Fonts