+++ packages/vue3/bili.config.js
@@ -1,0 +1,4 @@
+import pkg from './package.json'
+import biliConfig from '../../bili.rootconfig'
+
+module.exports = biliConfig(pkg)