+++ packages/vue3/src/utils.js
@@ -1,4 +1,4 @@
-import { xyzTransitionClasses, getXyzAnimationHook } from '../../../../utils'
+import { xyzTransitionClasses, getXyzAnimationHook } from '../../../utils'
 
 export function getXyzTransitionData(data) {
 	const { appear, duration } = data || {}