+++ packages/vue3/CHANGELOG.md
@@ -5,7 +5,10 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 # [0.2.0-alpha.0](https://github.com/ingram-projects/animxyz/compare/v0.1.2...v0.2.0-alpha.0) (2020-12-29)
 
-**Note:** Version bump only for package @animxyz/vue2-examples
+
+### Features
+
+* **vue:** added vue 3 support ([f0f1aa2](https://github.com/ingram-projects/animxyz/commit/f0f1aa27a85afd3b025ed6ce45f6e38974468e36))
 
 
 
@@ -13,4 +16,4 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [0.1.2](https://github.com/ingram-projects/animxyz/compare/v0.1.1...v0.1.2) (2020-12-29)
 
-**Note:** Version bump only for package @animxyz/vue-examples
+**Note:** Version bump only for package @animxyz/vue3